+++ src/projectM-sdl/pmSDL.cpp
@@ -585,7 +585,7 @@ void projectMSDL::renderTexture() {
 }
 
 void projectMSDL::presetSwitchedEvent(bool isHardCut, size_t index) const {
-    std::string presetName = getPresetName(index);
+    std::string presetName = getPresetName((uint)index);
     SDL_LogInfo(SDL_LOG_CATEGORY_APPLICATION, "Displaying preset: %s\n", presetName.c_str());
     
     std::string newTitle = "projectM ➫ " + presetName;
